requirements.txt
mujoco>=3.0
gymnasium>=0.29
stable-baselines3[extra]>=2.1
robot_descriptions
scipy
numpy
matplotlib
pyyaml